+++ tests/testthat/test-provider-snowflake.R
@@ -27,12 +27,10 @@ test_that("defaults are reported", {
 })
 
 test_that("all tool variations work", {
-  # Snowflake models don't support tool calling.
-  #
-  # test_tools_simple(chat_snowflake)
-  # test_tools_async(chat_snowflake)
-  # test_tools_parallel(chat_snowflake)
-  # test_tools_sequential(chat_snowflake, total_calls = 6)
+  test_tools_simple(chat_snowflake)
+  test_tools_async(chat_snowflake)
+  test_tools_parallel(chat_snowflake, total_calls = 6)
+  test_tools_sequential(chat_snowflake, total_calls = 6)
 })
 
 test_that("can extract data", {